refactor moviess in RateViewController to new class
