+++ ios_ui/MovieRec.xcworkspace/xcuserdata/yujin.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,12 +10,28 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "MovieRec/RateViewController.swift"
-            timestampString = "524718606.836342"
+            timestampString = "524815840.45909"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "32"
-            endingLineNumber = "32"
+            startingLineNumber = "33"
+            endingLineNumber = "33"
             landmarkName = "rateSkipButton(_:)"
+            landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "MovieRec/RateViewController.swift"
+            timestampString = "524815898.839531"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "95"
+            endingLineNumber = "95"
+            landmarkName = "processRating(ratingType:)"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
